+++ app/profile/edit/profileForms.tsx
@@ -1,0 +1,320 @@
+import React, { MouseEventHandler, useEffect, useState } from "react";
+import Image from "next/image";
+import { PhotoIcon } from "@heroicons/react/24/solid";
+import { useS3Upload } from "next-s3-upload";
+import UpdateDialog from "@/app/profile/edit/updateDialog";
+import { useRouter } from "next/navigation";
+
+export default function ProfileForms() {
+  const [profile, setProfile] = useState({
+    bio: "",
+    profilePicture: "",
+    bannerPicture: "",
+    displayName: "",
+  });
+
+  const router = useRouter();
+
+  const { uploadToS3 } = useS3Upload();
+  const [isDialogOpen, setDialogOpen] = useState(false);
+
+  const handleFileChange = async (file: File) => {
+    const { url } = await uploadToS3(file);
+    return url;
+  };
+
+  const handleSubmit: MouseEventHandler<HTMLButtonElement> = (e) => {
+    e.preventDefault();
+    fetch("/api/profile", {
+      method: "POST",
+      headers: {
+        "Content-Type": "application/json",
+      },
+      body: JSON.stringify(profile),
+    }).then((res) => {
+      if (res.ok) {
+        setDialogOpen(true);
+      } else {
+        alert("Profile update failed");
+      }
+    });
+  };
+
+  useEffect(() => {
+    fetch("/api/profile")
+      .then((res) => res.json())
+      .then((data) => {
+        setProfile({
+          bio: data.bio,
+          profilePicture: data.profilePicture,
+          bannerPicture: data.bannerPicture,
+          displayName: data.displayName,
+        });
+      });
+  }, []);
+
+  return (
+    <div className="mx-auto max-w-7xl px-2 sm:px-4 lg:divide-y lg:divide-gray-200 lg:px-8">
+      <div className="relative flex h-16 justify-between">
+        <form className="p-12">
+          <div className="space-y-12">
+            <div className="grid grid-cols-1 gap-x-8 gap-y-10 border-b border-gray-900/10 pb-12 md:grid-cols-3">
+              <div>
+                <h2 className="text-base font-semibold leading-7 text-gray-900">
+                  Profile
+                </h2>
+                <p className="mt-1 text-sm leading-6 text-gray-600">
+                  This information will be displayed publicly so be careful what
+                  you share.
+                </p>
+              </div>
+
+              <UpdateDialog
+                isOpen={isDialogOpen}
+                setIsOpen={setDialogOpen}
+                router={router}
+              />
+
+              <div className="grid max-w-2xl grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6 md:col-span-2">
+                <div className="col-span-full">
+                  <label
+                    htmlFor="about"
+                    className="block text-sm font-medium leading-6 text-gray-900"
+                  >
+                    About
+                  </label>
+                  <div className="mt-2">
+                    <textarea
+                      id="about"
+                      name="about"
+                      rows={3}
+                      value={profile.bio}
+                      onChange={(e) => {
+                        setProfile({ ...profile, bio: e.target.value });
+                      }}
+                      className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                    />
+                  </div>
+                  <p className="mt-3 text-sm leading-6 text-gray-600">
+                    Write a few sentences about yourself.
+                  </p>
+                </div>
+
+                <div className="col-span-full">
+                  <label
+                    htmlFor="photo"
+                    className="block text-sm font-medium leading-6 text-gray-900"
+                  >
+                    Photo
+                  </label>
+                  <div className="mt-2 flex items-center gap-x-3">
+                    {profile.profilePicture ? (
+                      <Image
+                        src={profile.profilePicture}
+                        width={42}
+                        height={42}
+                        alt={"profile picture"}
+                      />
+                    ) : (
+                      <Image
+                        src={
+                          "https://upload.wikimedia.org/wikipedia/commons/2/2c/Default_pfp.svg"
+                        }
+                        width={42}
+                        height={42}
+                        alt={"profile picture"}
+                      />
+                    )}
+
+                    <button
+                      type="button"
+                      className="rounded-md bg-white px-2.5 py-1.5 text-sm font-semibold text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 hover:bg-gray-50"
+                    >
+                      <label
+                        htmlFor="profile-file-upload"
+                        className="relative cursor-pointer rounded-md bg-white font-semibold text-indigo-600 hover:text-indigo-500"
+                      >
+                        <span>Change</span>
+                        <input
+                          id="profile-file-upload"
+                          name="profile-file-upload"
+                          type="file"
+                          onChange={(e) => {
+                            const file = e.target.files?.[0];
+
+                            if (file) {
+                              handleFileChange(file).then((r) => {
+                                setProfile({
+                                  ...profile,
+                                  profilePicture: r,
+                                });
+                              });
+                            }
+                          }}
+                          className="sr-only"
+                        />
+                      </label>
+                    </button>
+                  </div>
+                </div>
+
+                <div className="col-span-full">
+                  <label
+                    htmlFor="cover-photo"
+                    className="block text-sm font-medium leading-6 text-gray-900"
+                  >
+                    Cover photo
+                  </label>
+                  {
+                    // Display the current banner picture if available
+                    profile.bannerPicture && (
+                      <div className="mt-2 flex justify-center rounded-lg border border-gray-900/25">
+                        <Image
+                          src={profile.bannerPicture}
+                          alt="Cover"
+                          width={0}
+                          height={0}
+                          sizes="100vw"
+                          style={{ width: "auto", height: "auto" }} // optional
+                          className="rounded-lg"
+                        />
+                      </div>
+                    )
+                  }
+                  <div className="mt-2 flex justify-center rounded-lg border border-dashed border-gray-900/25 px-6 py-10">
+                    <div className="text-center">
+                      <PhotoIcon
+                        className="mx-auto h-12 w-12 text-gray-300"
+                        aria-hidden="true"
+                      />
+                      <div className="mt-4 flex text-sm leading-6 text-gray-600">
+                        <label
+                          htmlFor="banner-file-upload"
+                          className="relative cursor-pointer rounded-md bg-white font-semibold text-indigo-600 focus-within:outline-none focus-within:ring-2 focus-within:ring-indigo-600 focus-within:ring-offset-2 hover:text-indigo-500"
+                        >
+                          <span>Upload a file</span>
+                          <input
+                            id="banner-file-upload"
+                            name="banner-file-upload"
+                            type="file"
+                            onChange={(e) => {
+                              const file = e.target.files?.[0];
+
+                              if (file) {
+                                handleFileChange(file).then((r) => {
+                                  setProfile({
+                                    ...profile,
+                                    bannerPicture: r,
+                                  });
+                                });
+                              }
+                            }}
+                            className="sr-only"
+                          />
+                        </label>
+                        <p className="pl-1">or drag and drop</p>
+                      </div>
+                      <p className="text-xs leading-5 text-gray-600">
+                        PNG, JPG, GIF up to 10MB
+                      </p>
+                    </div>
+                  </div>
+                </div>
+              </div>
+            </div>
+
+            <div className="grid grid-cols-1 gap-x-8 gap-y-10 border-b border-gray-900/10 pb-12 md:grid-cols-3">
+              <div>
+                <h2 className="text-base font-semibold leading-7 text-gray-900">
+                  Personal Information
+                </h2>
+                <p className="mt-1 text-sm leading-6 text-gray-600">
+                  Use a permanent address where you can receive mail.
+                </p>
+              </div>
+
+              <div className="grid max-w-2xl grid-cols-1 gap-x-6 gap-y-8 sm:grid-cols-6 md:col-span-2">
+                <div className="sm:col-span-3">
+                  <label
+                    htmlFor="first-name"
+                    className="block text-sm font-medium leading-6 text-gray-900"
+                  >
+                    Display Name
+                  </label>
+                  <div className="mt-2">
+                    <input
+                      type="text"
+                      name="first-name"
+                      id="first-name"
+                      autoComplete="given-name"
+                      value={profile.displayName}
+                      onChange={(e) => {
+                        setProfile({ ...profile, displayName: e.target.value });
+                      }}
+                      className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                    />
+                  </div>
+                </div>
+
+                <div className="sm:col-span-3">
+                  <label
+                    htmlFor="country"
+                    className="block text-sm font-medium leading-6 text-gray-900"
+                  >
+                    Country
+                  </label>
+                  <div className="mt-2">
+                    <select
+                      id="country"
+                      name="country"
+                      autoComplete="country-name"
+                      className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:max-w-xs sm:text-sm sm:leading-6"
+                    >
+                      <option>United States</option>
+                      <option>Canada</option>
+                      <option>Mexico</option>
+                    </select>
+                  </div>
+                </div>
+
+                <div className="sm:col-span-2">
+                  <label
+                    htmlFor="postal-code"
+                    className="block text-sm font-medium leading-6 text-gray-900"
+                  >
+                    SSN
+                  </label>
+                  <div className="mt-2">
+                    <input
+                      type="text"
+                      name="postal-code"
+                      id="postal-code"
+                      autoComplete="postal-code"
+                      className="block w-full rounded-md border-0 py-1.5 text-gray-900 shadow-sm ring-1 ring-inset ring-gray-300 placeholder:text-gray-400 focus:ring-2 focus:ring-inset focus:ring-indigo-600 sm:text-sm sm:leading-6"
+                    />
+                  </div>
+                </div>
+              </div>
+            </div>
+          </div>
+
+          <div className="mt-6 flex items-center justify-end gap-x-6">
+            <button
+              type="button"
+              className="text-sm font-semibold leading-6 text-gray-900"
+            >
+              Cancel
+            </button>
+            <button
+              type="submit"
+              onClick={handleSubmit}
+              className="rounded-md bg-indigo-600 px-3 py-2 text-sm font-semibold text-white shadow-sm hover:bg-indigo-500 focus-visible:outline focus-visible:outline-2 focus-visible:outline-offset-2 focus-visible:outline-indigo-600"
+            >
+              Save
+            </button>
+          </div>
+        </form>
+      </div>
+    </div>
+  );
+}
