+++ middleware.js
@@ -1,3 +1,5 @@
 export { default } from "next-auth/middleware";
 
-export const config = { matcher: ["/profile"] };
+export const config = {
+  matcher: ["/profile", "/profile/edit"],
+};
